\name{coryphanthaA}
\alias{coryphanthaA}
\alias{coryphanthaB}
\alias{coryphanthaC}
\docType{data}
\title{
Transition matrices of three Coryphantha robbinsorum populations
}
\description{
Transition matrices of three  \emph{Coryphantha robbinsorum} populations.
}
\usage{data(coryphanthaA)}
\format{
  3x3 matrix.
}

\source{
  Although the original data are from Schmalzel et al (1995), these matrices have
  been transcribed from  Gurevitch et al. (2002). 
}
\references{

  Gurevitch, J., Scheiner, S. M. and Fox, G. A. 2002.
  \emph{The Ecology of Plants}. Sinauer.

  Schmalzel, R.J., F.W. Reichenbacher, and S. Rutman. 1995.
  Demographic study of the rare \emph{Coryphantha robbinsorum} (Cactaceae) 
  in southeastern Arizona. \emph{Madrono} 42:332-348.
}
\examples{

data(coryphanthaA)

}
\keyword{datasets}
